run on windows
make it run on windows
1. long -> int64_t
2. compiler command change

+++ geotransformer/extensions/cpu/radius_neighbors/radius_neighbors.cpp
@@ -35,10 +35,10 @@ at::Tensor radius_neighbors(
     reinterpret_cast<PointXYZ*>(s_points.data_ptr<float>()) + total_s_points
   );
   std::vector<long> vec_q_lengths = std::vector<long>(
-    q_lengths.data_ptr<long>(), q_lengths.data_ptr<long>() + batch_size
+    q_lengths.data_ptr<int64_t>(), q_lengths.data_ptr<int64_t>() + batch_size
   );
   std::vector<long> vec_s_lengths = std::vector<long>(
-    s_lengths.data_ptr<long>(), s_lengths.data_ptr<long>() + batch_size
+    s_lengths.data_ptr<int64_t>(), s_lengths.data_ptr<int64_t>() + batch_size
   );
   std::vector<long> vec_neighbor_indices;
 
@@ -54,12 +54,12 @@ at::Tensor radius_neighbors(
   std::size_t max_neighbors = vec_neighbor_indices.size() / total_q_points;
 
   at::Tensor neighbor_indices = torch::zeros(
-    {total_q_points, max_neighbors},
+    {static_cast<int64_t>(total_q_points), static_cast<int64_t>(max_neighbors)},
     at::device(q_points.device()).dtype(at::ScalarType::Long)
   );
 
   std::memcpy(
-    neighbor_indices.data_ptr<long>(),
+    neighbor_indices.data_ptr<int64_t>(),
     vec_neighbor_indices.data(),
     sizeof(long) * total_q_points * max_neighbors
   );
